requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-MySQLdb==0.2.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
python-dotenv==0.18.0
requests==2.25.1
urllib3==1.26.6
Werkzeug==2.0.1
